Fix build failure on latest nightly
