+++ tests/ui/forget-semicolon.stderr
@@ -4,7 +4,7 @@ error[E0308]: mismatched types
 8 |     Some(()) //~ ERROR mismatched types
   |     ^^^^^^^^- help: try adding a semicolon: `;`
   |     |
-  |     expected (), found enum `std::option::Option`
+  |     expected `()`, found enum `std::option::Option`
   |
-  = note: expected type `()`
-             found type `std::option::Option<()>`
+  = note: expected unit type `()`
+                  found enum `std::option::Option<()>`
