+++ tests/ui/type-error.stderr
@@ -2,7 +2,6 @@ error[E0308]: mismatched types
  --> $DIR/type-error.rs:7:18
   |
 7 |     let a: i32 = "a"; //~ ERROR mismatched types
-  |                  ^^^ expected i32, found reference
-  |
-  = note: expected type `i32`
-             found type `&'static str`
+  |            ---   ^^^ expected `i32`, found `&str`
+  |            |
+  |            expected due to this
